+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## next
+## 1.1.0 (Oct 1, 2021)
 
 - Add 2 new options: `convertTypes` (#123) and `resize` (#130).
 - Ignore the `strict` option when the `maxWidth/Height` option is set and its value is less than the natural width/height of the image (#134).